+++ examples/terraform/bootkube-install/provider.tf
@@ -1,9 +1,14 @@
 // Configure the matchbox provider
 provider "matchbox" {
+  version = "0.2.2"
   endpoint    = "${var.matchbox_rpc_endpoint}"
   client_cert = "${file("~/.matchbox/client.crt")}"
   client_key  = "${file("~/.matchbox/client.key")}"
   ca          = "${file("~/.matchbox/ca.crt")}"
+}
+
+provider "ct" {
+  version = "0.3.0"
 }
 
 provider "local" {